requirements.txt
feedparser==6.0.11
trafilatura==1.12.0
openai>=1.30.0
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.1
sqlalchemy==2.0.30
